requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
PyJWT==2.6.0
pymongo==4.3.3
python-dotenv==1.0.0
Werkzeug==2.2.3
